+++ backend/src/main/java/ca/metricalsky/winston/entity/view/AuthorDetails.java
@@ -1,0 +1,14 @@
+package ca.metricalsky.winston.entity.view;
+
+import ca.metricalsky.winston.entity.Author;
+
+public interface AuthorDetails {
+
+    Author getAuthor();
+
+    Long getCommentedVideos();
+
+    Long getTotalComments();
+
+    Long getTotalReplies();
+}