+++ frontend/src/model/authors/AuthorDto.ts
@@ -3,4 +3,11 @@ export type AuthorDto = {
     displayName: string;
     channelUrl: string;
     profileImageUrl: string;
+    statistics?: AuthorStatistics;
+}
+
+type AuthorStatistics = {
+    commentedVideos: number;
+    totalComments: number;
+    totalReplies: number;
 }